DTO-266: 프로필 변경 API 추가

+++ src/main/java/com/dto/way/member/web/dto/MemberResponseDTO.java
@@ -46,9 +46,9 @@ public class MemberResponseDTO {
 
         private String nickname;
 
-        private Long dailyCount;
+        private Integer dailyCount;
 
-        private Long historyCount;
+        private Integer historyCount;
 
         private Long followingCount;
